requirements.txt
aiofiles==24.1.0
argcomplete==3.4.0
attrs==23.2.1.dev0
autocommand==2.2.2
btrfsutil==6.10
build==1.2.2
CacheControl==0.14.0
cachy==0.3.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
cryptography==42.0.8
cssselect==1.2.0
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
dulwich==0.22.2
et-xmlfile==1.1.0
fastjsonschema==2.20.0
filelock==3.13.3
html2text==2024.2.26
html5lib==1.1
idna==3.8
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.2
jaraco.text==4.0.0
jeepney==0.8.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==25.2.1
lark==1.1.9
lockfile==0.12.2
lxml==5.3.0
Mako==1.3.5.dev0
Markdown==3.7
MarkupSafe==2.1.5
meson==1.5.2
more-itertools==10.3.0
msgpack==1.0.5
mysql-connector-python==9.1.0
mysqlx-connector-python==9.1.0
numpy==2.1.1
openpyxl==3.0.10
ordered-set==4.1.0
packaging==24.1
pandas==2.2.2
pexpect==4.9.0
pillow==10.4.0
pipx==1.6.0
pkginfo==1.10.0
platformdirs==4.3.6
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pycairo==1.27.0
pycparser==2.22
PyGObject==3.50.0
pyprland==2.4.0
pyproject_hooks==1.2.0
PyQt6==6.8.0.dev2410141303
PyQt6-WebEngine==6.7.0
PyQt6_sip==13.8.0
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
pywal==3.3.0
PyYAML==6.0.2
qutebrowser==3.2.1
rapidfuzz==3.6.2
referencing==0.35.1
requests==2.32.3
requests-futures==1.0.1
requests-toolbelt==1.0.0
rpds-py==0.19.0
SecretStorage==3.3.3
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
stem==1.8.2
tomli==2.0.1
tomlkit==0.13.2
torrequest==0.1.0
tqdm==4.66.5
trove-classifiers==2024.9.17
typing_extensions==4.12.2
tzdata==2024.1
ueberzug==18.2.3
urllib3==1.26.20
userpath==1.9.2
validate-pyproject==0.20.2
virtualenv==20.26.2
webencodings==0.5.1
yt-dlp==2024.10.7
